Extracted ThemePicker to its own component
