+++ screens/profile/SettingsScreen.tsx
@@ -1,16 +1,16 @@
-import { Ionicons, MaterialCommunityIcons } from "@expo/vector-icons";
+import { Ionicons } from "@expo/vector-icons";
 import { NativeStackScreenProps } from "@react-navigation/native-stack";
 import Constants from "expo-constants";
 import { signOut } from "firebase/auth";
 import React from "react";
-import { ColorSchemeName, Pressable, View } from "react-native";
-import { Button, RadioButton, Surface, Text, useTheme } from "react-native-paper";
+import { Pressable } from "react-native";
+import { Button, Text, useTheme } from "react-native-paper";
 import styled from "styled-components/native";
+import ThemePicker from "../../components/common/ThemePicker";
 import { auth } from "../../config/firebase";
 import { useUtils } from "../../hooks/useUtils";
 import { RootStackParams } from "../../navigation/RootStackNavigator";
-import { useAppDispatch, useAppSelector } from "../../store/store";
-import { selectTheme, set } from "../../store/theme/themeSlice";
+import { useAppDispatch } from "../../store/store";
 import { logout } from "../../store/user/userSlice";
 
 type Props = NativeStackScreenProps<RootStackParams>;
@@ -18,7 +18,6 @@ type Props = NativeStackScreenProps<RootStackParams>;
 const SettingsScreen = ({ navigation }: Props) => {
   const { colors } = useTheme();
   const dispatch = useAppDispatch();
-  const theme = useAppSelector(selectTheme);
   const { resetStore } = useUtils();
 
   function handleSignOut() {
@@ -41,39 +40,7 @@ const SettingsScreen = ({ navigation }: Props) => {
         </Text>
       </HeaderContainer>
       <Container>
-        <ModalContent elevation={0}>
-          <View>
-            <Text style={{ marginLeft: 10 }}>
-              darkmode <MaterialCommunityIcons name="theme-light-dark" size={15} color={colors.primary} />
-            </Text>
-
-            <RadioButton.Group onValueChange={(theme) => dispatch(set(theme))} value={theme.theme as NonNullable<ColorSchemeName>}>
-              <Content>
-                <RadioButtonContainer onPress={() => dispatch(set("dark"))}>
-                  <Text variant="bodyMedium">På</Text>
-                  <RadioButton.Android value="dark" />
-                </RadioButtonContainer>
-              </Content>
-
-              <Content>
-                <RadioButtonContainer onPress={() => dispatch(set("light"))}>
-                  <Text variant="bodyMedium">Av</Text>
-                  <RadioButton.Android value="light" />
-                </RadioButtonContainer>
-              </Content>
-
-              <Content>
-                <RadioButtonContainer onPress={() => dispatch(set("system"))}>
-                  <View>
-                    <Text variant="bodyMedium">System</Text>
-                    <Text variant="bodySmall">Vi anpassar applikationen baserat på din enhets systeminställningar.</Text>
-                  </View>
-                  <RadioButton.Android value="system" />
-                </RadioButtonContainer>
-              </Content>
-            </RadioButton.Group>
-          </View>
-        </ModalContent>
+        <ThemePicker />
         <Button style={{ marginBottom: 20 }} mode="contained" onPress={handleSignOut}>
           Logga ut
         </Button>
@@ -86,21 +53,6 @@ export default SettingsScreen;
 
 const Container = styled.View`
   flex: 1;
-  align-items: center;
-  justify-content: space-between;
-`;
-const Content = styled(Surface)`
-  margin: 5px;
-  padding: 10px 20px;
-  border-radius: 10px;
-`;
-const ModalContent = styled(Surface)`
-  padding: 20px;
-  margin-top: 10px;
-`;
-
-const RadioButtonContainer = styled.TouchableOpacity`
-  flex-direction: row;
   align-items: center;
   justify-content: space-between;
 `;
